requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
python-dotenv
azure-cosmos
requests  # Optional
pydantic  # Optional but useful for JSON validation
pymongo